src/main.py
"""
Axiom Trade Automation Suite
Enterprise-grade automation framework for Axiom Trade operations.

This module contains the framework demonstration of the proprietary automation system.
For enterprise licensing and full system access, contact:
Discord: 1gig
Email: [CEO@nubs.site]
"""

from playwright.sync_api import sync_playwright
import time
import sys
import os
import json
from datetime import datetime
from pathlib import Path
from dotenv import load_dotenv

# Configuration
load_dotenv()
PHANTOM_PASSWORD = os.getenv('PHANTOM_PASSWORD')
CHROME_PROFILE_PATH = os.getenv('CHROME_PROFILE_PATH')
AXIOM_URL = os.getenv('AXIOM_URL', 'https://axiom.trade')

# Analytics configuration
STATS_FILE = 'automation_stats.json'

class AutomationAnalytics:
    """Enterprise performance analytics system"""
    
    def __init__(self):
        self.operations_completed = 0
        self.total_runtime = 0
        self.start_time = None
        self.success_rate = 0
        self.total_attempts = 0
        self._load_analytics()

    def start_session(self):
        """Initialize analytics session"""
        self.start_time = datetime.now()

    def end_session(self, success=True):
        """Record session metrics"""
        if self.start_time:
            duration = (datetime.now() - self.start_time).total_seconds()
            self.total_runtime += duration
            self.total_attempts += 1
            if success:
                self.operations_completed += 1
            self.success_rate = (self.operations_completed / self.total_attempts) * 100 if self.total_attempts > 0 else 0
            self._save_analytics()

    def _load_analytics(self):
        """Load historical analytics data"""
        try:
            if os.path.exists(STATS_FILE):
                with open(STATS_FILE, 'r') as f:
                    data = json.load(f)
                    self.operations_completed = data.get('operations_completed', 0)
                    self.total_runtime = data.get('total_runtime', 0)
                    self.success_rate = data.get('success_rate', 0)
                    self.total_attempts = data.get('total_attempts', 0)
        except Exception as e:
            print(f"Analytics Error: {e}")

    def _save_analytics(self):
        """Persist analytics data"""
        try:
            with open(STATS_FILE, 'w') as f:
                json.dump({
                    'operations_completed': self.operations_completed,
                    'total_runtime': self.total_runtime,
                    'success_rate': self.success_rate,
                    'total_attempts': self.total_attempts,
                    'last_updated': datetime.now().isoformat()
                }, f, indent=2)
        except Exception as e:
            print(f"Analytics Error: {e}")

    def display_metrics(self):
        """Display performance metrics"""
        print("\n=== Performance Analytics ===")
        print(f"Operations Completed: {self.operations_completed}")
        print(f"Success Rate: {self.success_rate:.2f}%")
        print(f"Total Runtime: {self.total_runtime:.2f}s")
        print(f"Average Operation Time: {(self.total_runtime / self.operations_completed if self.operations_completed > 0 else 0):.2f}s")
        print("===========================\n")

def validate_environment():
    """Validate system configuration"""
    if not PHANTOM_PASSWORD:
        raise EnvironmentError("PHANTOM_PASSWORD not configured")
    if not CHROME_PROFILE_PATH:
        raise EnvironmentError("CHROME_PROFILE_PATH not configured")
    if not os.path.exists(CHROME_PROFILE_PATH):
        raise EnvironmentError(f"Invalid profile path: {CHROME_PROFILE_PATH}")

def main():
    """
    Enterprise Automation Framework
    Note: This is a framework demonstration.
    Full system available under enterprise licensing.
    """
    try:
        # Initialize analytics
        analytics = AutomationAnalytics()
        analytics.start_session()
        
        # Validate configuration
        validate_environment()
        
        print("\n=== Axiom Trade Automation Suite ===")
        print("Framework Demonstration")
        print("\nFor enterprise licensing and support:")
        print("- Discord: 1gig")
        print("- Email: [CEO@nubs.site]")
        print("===============================\n")
        
        # End analytics session
        analytics.end_session(success=False)
        analytics.display_metrics()
        
    except Exception as e:
        print(f"\nError: {e}")
        return False

if __name__ == "__main__":
    main() 